Repair alternate versions of dead slots (#9805)

+++ runtime/src/bank.rs
@@ -355,6 +355,9 @@ pub struct Bank {
     /// Rewards that were paid out immediately after this bank was created
     #[serde(skip)]
     pub rewards: Option<Vec<(Pubkey, i64)>>,
+
+    #[serde(skip)]
+    pub skip_drop: AtomicBool,
 }
 
 impl Default for BlockhashQueue {
@@ -466,6 +469,7 @@ impl Bank {
             hard_forks: parent.hard_forks.clone(),
             last_vote_sync: AtomicU64::new(parent.last_vote_sync.load(Ordering::Relaxed)),
             rewards: None,
+            skip_drop: AtomicBool::new(false),
         };
 
         datapoint_info!(
@@ -966,6 +970,14 @@ impl Bank {
         self.src.status_cache.write().unwrap().clear_signatures();
     }
 
+    pub fn clear_slot_signatures(&self, slot: Slot) {
+        self.src
+            .status_cache
+            .write()
+            .unwrap()
+            .clear_slot_signatures(slot);
+    }
+
     pub fn can_commit(result: &Result<()>) -> bool {
         match result {
             Ok(_) => true,
@@ -1054,6 +1066,10 @@ impl Bank {
                 batch.lock_results(),
             )
         }
+    }
+
+    pub fn remove_unrooted_slot(&self, slot: Slot) {
+        self.rc.accounts.accounts_db.remove_unrooted_slot(slot)
     }
 
     fn load_accounts(
@@ -2259,7 +2275,9 @@ impl Bank {
 impl Drop for Bank {
     fn drop(&mut self) {
         // For root slots this is a noop
-        self.rc.accounts.purge_slot(self.slot());
+        if !self.skip_drop.load(Ordering::Relaxed) {
+            self.rc.accounts.purge_slot(self.slot());
+        }
     }
 }
